+++ src/main/java/ru/vyarus/dropwizard/guice/module/jersey/support/JerseyComponentProvider.java
@@ -1,6 +1,7 @@
 package ru.vyarus.dropwizard.guice.module.jersey.support;
 
 import com.google.inject.Injector;
+import org.glassfish.hk2.api.ProxyCtl;
 import org.glassfish.hk2.api.ServiceLocator;
 
 import javax.inject.Provider;
@@ -25,7 +26,18 @@ public class JerseyComponentProvider<T> implements Provider<T> {
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public T get() {
-        return injector.get().getInstance(ServiceLocator.class).getService(type);
+        // hk2 by default proxy instances to delay actual instance creation, which could harm guice scopes logic
+        // for example: if guice request scope transfer used ServletScopes.transferRequest and we try to obtain
+        // it will try to use proxy instance in separate thread which will perform hk2 checks for request scope
+        // and fail. Instead, we always resolve actual instance and let guice properly control scoping
+        final T res = injector.get().getInstance(ServiceLocator.class).getService(type);
+        return res instanceof ProxyCtl ? (T) ((ProxyCtl) res).__make() : res;
+    }
+
+    @Override
+    public String toString() {
+        return "JerseyComponentProvider for " + type;
     }
 }